Port tests for old utils file, delete
